+++ crates/solverforge-solver/src/phase/vnd/phase.rs
@@ -234,15 +234,15 @@ mod tests {
     };
     use solverforge_core::score::SimpleScore;
 
-    type NQueensMove = ChangeMove<NQueensSolution, i32>;
+    type NQueensMove = ChangeMove<NQueensSolution, i64>;
 
     fn create_move_selector(
-        values: Vec<i32>,
+        values: Vec<i64>,
     ) -> ChangeMoveSelector<
         NQueensSolution,
-        i32,
+        i64,
         crate::heuristic::selector::FromSolutionEntitySelector,
-        crate::heuristic::selector::StaticTypedValueSelector<NQueensSolution, i32>,
+        crate::heuristic::selector::StaticTypedValueSelector<NQueensSolution, i64>,
     > {
         ChangeMoveSelector::simple(get_queen_row, set_queen_row, 0, "row", values)
     }
@@ -255,7 +255,7 @@ mod tests {
         let initial_score = solver_scope.calculate_score();
         assert!(initial_score < SimpleScore::of(0));
 
-        let values: Vec<i32> = (0..4).collect();
+        let values: Vec<i64> = (0..4).collect();
         let mut phase: VndPhase<_, NQueensMove> = VndPhase::new((
             create_move_selector(values.clone()),
             create_move_selector(values),
@@ -274,7 +274,7 @@ mod tests {
 
         let initial_score = solver_scope.calculate_score();
 
-        let values: Vec<i32> = (0..4).collect();
+        let values: Vec<i64> = (0..4).collect();
         let mut phase: VndPhase<_, NQueensMove> = VndPhase::new((create_move_selector(values),));
 
         phase.solve(&mut solver_scope);
